+++ src/icalendar/cal/calendar.py
@@ -472,6 +472,8 @@ Description:
 
         Raises:
             IncompleteComponent: If the content is not valid according to :rfc:`5545`.
+
+        .. warning:: As time progresses, we will be stricter with the validation.
         """
         calendar = cls()
         calendar.prodid = prodid