+++ src/icalendar/cal/event.py
@@ -403,6 +403,8 @@ class Event(Component):
 
         Raises:
             IncompleteComponent: If the content is not valid according to :rfc:`5545`.
+
+        .. warning:: As time progresses, we will be stricter with the validation.
         """
         event = super().new(dtstamp=dtstamp if dtstamp is not None else cls._utc_now())
         event.summary = summary