+++ src/icalendar/cal/free_busy.py
@@ -76,6 +76,7 @@ class FreeBusy(Component):
         Raises:
             IncompleteComponent: If the content is not valid according to :rfc:`5545`.
 
+        .. warning:: As time progresses, we will be stricter with the validation.
         """
         free_busy = super().new(
             dtstamp=dtstamp if dtstamp is not None else cls._utc_now()