+++ src/icalendar/cal/todo.py
@@ -275,6 +275,8 @@ class Todo(Component):
 
         Raises:
             IncompleteComponent: If the content is not valid according to :rfc:`5545`.
+
+        .. warning:: As time progresses, we will be stricter with the validation.
         """
         todo = super().new(dtstamp=dtstamp if dtstamp is not None else cls._utc_now())
         todo.summary = summary